refactor: move getFontKitFont and heightOfFontAtSize methods to common
