+++ packages/common/src/helper.ts
@@ -1,5 +1,7 @@
 import { z } from 'zod';
 import { Buffer } from 'buffer';
+import type { Font as FontKitFont } from 'fontkit';
+import * as fontkit from 'fontkit';
 import {
   Schema,
   Template,
@@ -279,4 +281,57 @@ export const checkGenerateProps = (data: unknown) => {
     migrateTemplate(data.template as Template);
   }
   checkProps(data, GeneratePropsSchema);
+};
+
+export const heightOfFontAtSize = (fontKitFont: FontKitFont, fontSize: number) => {
+  const { ascent, descent, bbox, unitsPerEm } = fontKitFont;
+
+  const scale = 1000 / unitsPerEm;
+  const yTop = (ascent || bbox.maxY) * scale;
+  const yBottom = (descent || bbox.minY) * scale;
+
+  let height = yTop - yBottom;
+  height -= Math.abs(descent * scale) || 0;
+
+  return (height / 1000) * fontSize;
+};
+
+
+const getFallbackFont = (font: Font) => {
+  const fallbackFontName = getFallbackFontName(font);
+  return font[fallbackFontName];
+};
+
+const getCacheKey = (fontName: string) => `getFontKitFont-${fontName}`;
+
+export const getFontKitFont = async (
+  fontName: string | undefined,
+  font: Font = getDefaultFont(),
+  _cache: Map<string | number, fontkit.Font>,
+) => {
+  const fntNm = fontName || getFallbackFontName(font);
+  const cacheKey = getCacheKey(fntNm);
+  if (_cache.has(cacheKey)) {
+    return _cache.get(cacheKey) as fontkit.Font;
+  }
+
+  const currentFont = font[fntNm] || getFallbackFont(font) || getDefaultFont()[DEFAULT_FONT_NAME];
+  let fontData = currentFont.data;
+  if (typeof fontData === 'string') {
+    fontData = fontData.startsWith('http')
+      ? await fetch(fontData).then((res) => res.arrayBuffer())
+      : b64toUint8Array(fontData);
+  }
+
+  // Convert fontData to Buffer if it's not already a Buffer
+  let fontDataBuffer: Buffer;
+  if (fontData instanceof Buffer) {
+    fontDataBuffer = fontData;
+  } else {
+    fontDataBuffer = Buffer.from(fontData as ArrayBufferLike);
+  }
+  const fontKitFont = fontkit.create(fontDataBuffer) as fontkit.Font;
+  _cache.set(cacheKey, fontKitFont);
+
+  return fontKitFont;
 };
