+++ packages/schemas/src/tables/classes.ts
@@ -1,6 +1,6 @@
-import { Font, mm2pt, pt2mm } from '@pdfme/common';
+import { Font, mm2pt, pt2mm, getFontKitFont } from '@pdfme/common';
 import type { Font as FontKitFont } from 'fontkit';
-import { splitTextToSize, getFontKitFont, widthOfTextAtSize } from '../text/helper.js';
+import { splitTextToSize, widthOfTextAtSize } from '../text/helper.js';
 import type { Styles, TableInput, Settings, Section, StylesProps } from './types.js';
 
 type ContentSettings = { body: Row[]; head: Row[]; columns: Column[] };
